Исправляет минимальную длину комментария
