+++ js/main.js
@@ -2,7 +2,6 @@ import {getThumbnails} from './thumbnails-rendering.js';
 import './form.js';
 import './scale.js';
 import './effects.js';
-import {updateSlider} from './slider.js';
 
 getThumbnails();
-updateSlider();
+
